+++ libft/srcs/ft_strcspn.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strcspn.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hasni <hasni@student.42.fr>                +#+  +:+       +#+        */
+/*   By: lutsiara <lutsiara@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/01/30 22:23:02 by hasni             #+#    #+#             */
-/*   Updated: 2020/01/30 22:25:38 by hasni            ###   ########.fr       */
+/*   Created: 2020/02/19 11:53:56 by lutsiara          #+#    #+#             */
+/*   Updated: 2020/02/19 11:57:47 by lutsiara         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strcspn(const char *s, const char *charset)
+unsigned long	ft_strcspn(const char *s, const char *charset)
 {
-	int	i;
+	unsigned long	i;
 
 	i = 0;
 	while (s[i] && !ft_charinset(s[i], charset))
